Moved handleAssessors method
