+++ src_ts/components/pages/assessments/assessment-tab-pages/details/external-individual.ts
@@ -7,7 +7,8 @@ import './external-individual-dialog';
 import {ExternalIndividualDialog} from './external-individual-dialog';
 import {connect} from 'pwa-helpers/connect-mixin';
 import {store, RootState} from '../../../../../redux/store';
-import {isJsonStrMatch, cloneDeep, handleAssessorsNoLongerAssignedToCurrentCountry} from '../../../../utils/utils';
+import {isJsonStrMatch, cloneDeep} from '../../../../utils/utils';
+import {handleAssessorsNoLongerAssignedToCurrentCountry} from '../../../../common/common-methods';
 import {EtoolsDropdownEl} from '@unicef-polymer/etools-dropdown/etools-dropdown';
 import {UnicefUser} from '../../../../../types/user-model';
 import {Assessor, AssessorTypes} from '../../../../../types/assessment';
