+++ src_ts/components/utils/utils.ts
@@ -1,5 +1,3 @@
-import {UnicefUser} from '../../types/user-model';
-
 export const isJsonStrMatch = (a: any, b: any) => {
   return JSON.stringify(a) === JSON.stringify(b);
 };
@@ -15,12 +13,3 @@ export const getFileNameFromURL = (url: string) => {
   // @ts-ignore
   return url.split('?').shift().split('/').pop();
 };
-
-export const handleAssessorsNoLongerAssignedToCurrentCountry = (assessors: UnicefUser[], assessorToCheck?: UnicefUser) => {
-  if (assessorToCheck && assessorToCheck.id && assessors) {
-    if (assessors.findIndex(user => user.id === assessorToCheck.id) < 0) {
-      assessors.push(assessorToCheck);
-      assessors.sort((a, b) => (a.name < b.name) ? -1 : 1);
-    }
-  }
-}
